chore: upgrade libp2p to 0.42.2
