+++ apps/hubble/package.json
@@ -29,7 +29,7 @@
   },
   "devDependencies": {
     "@faker-js/faker": "~7.6.0",
-    "@libp2p/interface-mocks": "7.0.3",
+    "@libp2p/interface-mocks": "^9.0.0",
     "@types/chance": "^1.1.3",
     "@types/node-cron": "^3.0.7",
     "@types/progress": "^2.0.5",
@@ -43,22 +43,22 @@
     "ts-mockito": "~2.6.1"
   },
   "dependencies": {
-    "@chainsafe/libp2p-gossipsub": "4.1.1",
-    "@chainsafe/libp2p-noise": "8.0.2",
+    "@chainsafe/libp2p-gossipsub": "6.1.0",
+    "@chainsafe/libp2p-noise": "^11.0.0 ",
     "@farcaster/protobufs": "0.1.3",
     "@farcaster/utils": "0.2.2",
     "@grpc/grpc-js": "~1.8.7",
     "@multiformats/multiaddr": "^11.0.0",
-    "@libp2p/interface-connection": "3.0.2",
-    "@libp2p/interface-peer-id": "1.1.2",
-    "@libp2p/mplex": "5.2.4",
-    "@libp2p/peer-id-factory": "1.0.19",
-    "@libp2p/pubsub-peer-discovery": "6.0.2",
-    "@libp2p/tcp": "3.1.2",
-    "@libp2p/utils": "3.0.2",
+    "@libp2p/interface-connection": "^3.0.2",
+    "@libp2p/interface-peer-id": "^2.0.0",
+    "@libp2p/mplex": "^7.0.0",
+    "@libp2p/peer-id-factory": "^2.0.0",
+    "@libp2p/pubsub-peer-discovery": "^8.0.0",
+    "@libp2p/tcp": "^6.0.0",
+    "@libp2p/utils": "^3.0.2",
     "commander": "~10.0.0",
     "ethers": "~5.7.2",
-    "libp2p": "0.39.5",
+    "libp2p": "0.42.2",
     "neverthrow": "~6.0.0",
     "node-cron": "~3.0.2",
     "pino": "~8.8.0",
